Add GenericTensorAttribute and trim down TensorAttribute.
