+++ searchlib/src/vespa/searchlib/tensor/tensor_attribute.cpp
@@ -4,7 +4,6 @@
 #include "tensor_attribute.h"
 #include <vespa/vespalib/tensor/default_tensor.h>
 #include <vespa/vespalib/tensor/tensor.h>
-#include "tensor_attribute_saver.h"
 
 using vespalib::eval::ValueType;
 using vespalib::tensor::Tensor;
@@ -21,20 +20,6 @@ constexpr uint32_t TENSOR_ATTRIBUTE_VERSION = 0;
 // minimum dead bytes in tensor attribute before consider compaction
 constexpr size_t DEAD_SLACK = 0x10000u;
 
-
-class TensorReader : public AttributeVector::ReaderBase
-{
-private:
-    FileReader<uint32_t> _tensorSizeReader;
-public:
-    TensorReader(AttributeVector &attr)
-        : AttributeVector::ReaderBase(attr),
-          _tensorSizeReader(*_datFile)
-    {
-    }
-    uint32_t getNextTensorSize() { return _tensorSizeReader.readHostOrder(); }
-    void readTensor(void *buf, size_t len) { _datFile->ReadBuf(buf, len); }
-};
 
 Tensor::UP
 createEmptyTensor(const TensorMapper *mapper)
@@ -55,13 +40,14 @@ shouldCreateMapper(const ValueType &tensorType)
 }
 
 TensorAttribute::TensorAttribute(const vespalib::stringref &baseFileName,
-                                 const Config &cfg)
+                                 const Config &cfg,
+                                 TensorStore &tensorStore)
     : NotImplementedAttribute(baseFileName, cfg),
       _refVector(cfg.getGrowStrategy().getDocsInitialCapacity(),
                  cfg.getGrowStrategy().getDocsGrowPercent(),
                  cfg.getGrowStrategy().getDocsGrowDelta(),
                  getGenerationHolder()),
-      _tensorStore(),
+      _tensorStore(tensorStore),
       _tensorMapper(),
       _compactGeneration(0)
 {
@@ -73,8 +59,6 @@ TensorAttribute::TensorAttribute(const vespalib::stringref &baseFileName,
 
 TensorAttribute::~TensorAttribute()
 {
-    getGenerationHolder().clearHoldLists();
-    _tensorStore.clearHoldLists();
 }
 
 
@@ -179,31 +163,18 @@ TensorAttribute::addDoc(DocId &docId)
 
 
 void
-TensorAttribute::setTensor(DocId docId, const Tensor &tensor)
+TensorAttribute::setTensorRef(DocId docId, RefType ref)
 {
     assert(docId < _refVector.size());
     updateUncommittedDocIdLimit(docId);
-    // TODO: Handle generic tensor attribute in a better way ?
-    RefType ref = _tensorStore.setTensor(
-            (_tensorMapper ? *_tensorMapper->map(tensor) : tensor));
     // TODO: validate if following fence is sufficient.
     std::atomic_thread_fence(std::memory_order_release);
     // TODO: Check if refVector must consist of std::atomic<RefType>
+    RefType oldRef(_refVector[docId]);
     _refVector[docId] = ref;
-}
-
-
-std::unique_ptr<Tensor>
-TensorAttribute::getTensor(DocId docId) const
-{
-    RefType ref;
-    if (docId < getCommittedDocIdLimit()) {
-        ref = _refVector[docId];
+    if (oldRef.valid()) {
+        _tensorStore.holdTensor(oldRef);
     }
-    if (!ref.valid()) {
-        return std::unique_ptr<Tensor>();
-    }
-    return _tensorStore.getTensor(ref);
 }
 
 Tensor::UP
@@ -238,32 +209,6 @@ TensorAttribute::onShrinkLidSpace()
 }
 
 
-bool
-TensorAttribute::onLoad()
-{
-    TensorReader tensorReader(*this);
-    if (!tensorReader.hasData()) {
-        return false;
-    }
-    setCreateSerialNum(tensorReader.getCreateSerialNum());
-    assert(tensorReader.getVersion() == TENSOR_ATTRIBUTE_VERSION);
-    uint32_t numDocs(tensorReader.getDocIdLimit());
-    _refVector.reset();
-    _refVector.unsafe_reserve(numDocs);
-    for (uint32_t lid = 0; lid < numDocs; ++lid) {
-        uint32_t tensorSize = tensorReader.getNextTensorSize();
-        auto raw = _tensorStore.allocRawBuffer(tensorSize);
-        if (tensorSize != 0) {
-            tensorReader.readTensor(raw.first, tensorSize);
-        }
-        _refVector.push_back(raw.second);
-    }
-    setNumDocs(numDocs);
-    setCommittedDocIdLimit(numDocs);
-    return true;
-}
-
-
 uint32_t
 TensorAttribute::getVersion() const
 {
@@ -277,18 +222,6 @@ TensorAttribute::getRefCopy() const
     uint32_t size = getCommittedDocIdLimit();
     assert(size <= _refVector.size());
     return RefCopyVector(&_refVector[0], &_refVector[0] + size);
-}
-
-std::unique_ptr<AttributeSaver>
-TensorAttribute::onInitSave()
-{
-    vespalib::GenerationHandler::Guard guard(getGenerationHandler().
-                                             takeGuard());
-    return std::make_unique<TensorAttributeSaver>
-        (std::move(guard),
-         this->createSaveTargetConfig(),
-         getRefCopy(),
-         _tensorStore);
 }
 
 }  // namespace search::attribute
